+++ operators/trim_three_point_edit.py
@@ -1,7 +1,6 @@
 import bpy
 from math import floor
 
-from .utils.find_strips_mouse import find_strips_mouse
 from .utils.get_mouse_view_coords import get_mouse_frame_and_channel
 
 
@@ -12,31 +11,29 @@ class TrimThreePointEdit(bpy.types.Operator):
     bl_options = {'REGISTER', 'UNDO'}
 
     side = bpy.props.EnumProperty(
-            items=[
-                ('left', 'Left', 'Left side'),
-                ('right', 'Right', 'Right side'),
-                ],
-            name="Trim side",
-            description="Side of the strip(s) to trim, either left or right",
-            default='left')
+        items=[
+            ('left', 'Left', 'Left side'),
+            ('right', 'Right', 'Right side'),
+        ],
+        name="Trim side",
+        description="Side of the strip(s) to trim, either left or right",
+        default='left')
 
     @classmethod
     def poll(cls, context):
-        return True
+        return context.sequences
 
     def invoke(self, context, event):
         frame, channel = get_mouse_frame_and_channel(event)
-
-        sequences = find_strips_mouse(frame, channel, select_linked=True)
-        if not sequences:
-            return {'CANCELLED'}
-
-        user_selection = bpy.context.selected_sequences
         bpy.ops.sequencer.select_all(action='DESELECT')
-        for s in sequences:
-            s.select = True
+        bpy.ops.power_sequencer.select_closest_to_mouse(
+            frame=frame, channel=channel)
+        if not bpy.context.selected_sequences:
+            bpy.ops.power_sequencer.select_strips_under_cursor()
+        return self.execute(context)
+
+    def execute(self, context):
+        if not bpy.context.selected_sequences:
+            return {'CANCELLED'}
         bpy.ops.power_sequencer.smart_snap(side=self.side)
-        # FIXME: remove the need for that crap
-        for s in user_selection:
-            s.select = True
         return {'FINISHED'}